requirements.txt
websocket-client==1.4.0
rich==12.5.1
sortedcontainers==2.4.0
requests==2.28.1
